+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/schema/RealmStorageTypeImpl.kt
@@ -28,6 +28,7 @@ internal object RealmStorageTypeImpl {
             io.realm.internal.interop.PropertyType.RLM_PROPERTY_TYPE_FLOAT -> RealmStorageType.FLOAT
             io.realm.internal.interop.PropertyType.RLM_PROPERTY_TYPE_DOUBLE -> RealmStorageType.DOUBLE
             io.realm.internal.interop.PropertyType.RLM_PROPERTY_TYPE_TIMESTAMP -> RealmStorageType.TIMESTAMP
+            io.realm.internal.interop.PropertyType.RLM_PROPERTY_TYPE_OBJECT_ID -> RealmStorageType.OBJECT_ID
             else -> error("Unknown storage type: $type")
         }
     }