+++ tools/sync_test_server/mongodb-realm-command-server.js
@@ -13,7 +13,10 @@
 var winston = require('winston'); //logging
 var http = require('http');
 var url = require('url');
-const fs = require('fs')
+const fs = require('fs');
+const { MongoClient, ObjectID } = require("mongodb");
+const mdb_uri =
+"mongodb://localhost:26000/?readPreference=primary&directConnection=true&ssl=false";
 
 const isPortAvailable = require('is-port-available');
 
@@ -111,6 +114,77 @@ function handleApplicationId(appName, req, resp) {
     }
 }
 
+async function handleMDBDocumentInsertRequest(clientReq, clientResp) {
+  try {
+    var url_parts = url.parse(clientReq.url, true);
+
+    const db_name = url_parts.query.db;
+    const collection = url_parts.query.collection;
+    const client = new MongoClient(mdb_uri);
+
+      switch(clientReq.method) {
+          case "PUT":
+              let data = '';
+              clientReq.on('data', chunk => {
+                data += chunk;
+              })
+              clientReq.on('end', async () => {
+                  const payload = JSON.parse(data);
+
+                  await client.connect();
+
+                  const database = client.db(db_name);
+                  const col = database.collection(collection);
+
+                  clientResp.writeHead(200, {'Content-Type': 'application/json'});
+                  var res = await col.insertOne(payload);
+                  clientResp.end(JSON.stringify(res));
+              });
+      }
+
+  } catch (err) {
+     console.error(err)
+     clientResp.writeHead(500, {'Content-Type': 'text/plain'});
+     clientResp.end();
+  } finally {
+    await client.close();
+  }
+}
+
+async function handleMDBDocumentQueryByIdRequest(clientReq, clientResp) {
+  try {
+    var url_parts = url.parse(clientReq.url, true);
+
+    const db_name = url_parts.query.db;
+    const collection = url_parts.query.collection;
+    const oid = ObjectID(url_parts.query.oid);
+    const client = new MongoClient(mdb_uri);
+
+      switch(clientReq.method) {
+          case "GET":
+              await client.connect();
+              const database = client.db(db_name);
+              const col = database.collection(collection);
+
+              var res = await col.findOne({_id: oid })
+              if (res == null) {
+                clientResp.writeHead(404, {'Content-Type': 'text/plain'});
+                clientResp.end();
+              } else {
+                 clientResp.writeHead(200, {'Content-Type': 'application/json'});
+                 clientResp.end(JSON.stringify(res));
+              }
+      }
+
+  } catch (err) {
+     console.error(err)
+     clientResp.writeHead(500, {'Content-Type': 'text/plain'});
+     clientResp.end();
+  } finally {
+    await client.close();
+  }
+}
+
 //Create and start the Http server
 const PORT = 8888;
 var applicationIds = {}  // Should be updated by the Docker setup script before any tests are run.
@@ -127,6 +201,10 @@ var server = http.createServer(function(req, resp) {
             handleWatcher(req, resp);
         } else if (req.url.includes('/forward-as-patch')) {
             handleForwardPatchRequest(req, resp);
+        } else if (req.url.includes("/insert-document")) {
+            handleMDBDocumentInsertRequest(req, resp);
+        } else if (req.url.includes("/query-document-by-id")) {
+            handleMDBDocumentQueryByIdRequest(req, resp);
         } else {
             handleUnknownEndPoint(req, resp);
         }